+++ src/rdkafka_op.c
@@ -136,7 +136,7 @@ void rd_kafka_op_app_reply (rd_kafka_q_t *rkq,
                             rd_kafka_op_type_t type,
                             rd_kafka_resp_err_t err,
                             int32_t version,
-                            void *payload, int len) {
+                            void *payload, size_t len) {
 	rd_kafka_op_t *rko;
 
         rko = rd_kafka_op_new(type);
@@ -229,7 +229,7 @@ void rd_kafka_q_op_err (rd_kafka_q_t *rkq, rd_kafka_op_type_t optype,
 void rd_kafka_op_app (rd_kafka_q_t *rkq, rd_kafka_op_type_t type,
                       int op_flags, rd_kafka_toppar_t *rktp,
                       rd_kafka_resp_err_t err,
-                      void *payload, int len,
+                      void *payload, size_t len,
                       void (*free_cb) (void *)) {
         rd_kafka_op_t *rko;
 
@@ -315,7 +315,7 @@ rd_kafka_op_t *rd_kafka_op_new_reply (rd_kafka_op_t *rko_orig) {
  */
 int rd_kafka_op_reply (rd_kafka_op_t *rko_orig,
                        rd_kafka_resp_err_t err,
-                       void *payload, int len, void (*free_cb) (void *)) {
+                       void *payload, size_t len, void (*free_cb) (void *)) {
         rd_kafka_op_t *rko;
 
         if (!rko_orig->rko_replyq)
